Restart cameraserver when qcamerasvr is starting
As mediaserver has been splitted, there might be a race condition
when starting cameraserver between it and qcamerasvr.

As cameraserver starts sometimes before qcamerasvr, the system
doesn't see a torch available for the Camera and blocks it.

Restart accordingly cameraserver when qcamerasvr is started.

I know that there are better ways of doing this, hoping that
this is a temporary solution.

Thanks to @xyyx for his initial fix:

nitrogen-project/android_device_lge_geehrc@0f728b1